extract last sub component
