+++ packages/perseus-editor/src/widgets/interaction-editor/index.ts
@@ -1,0 +1,3 @@
+import InteractionEditor from "./interaction-editor";
+
+export default InteractionEditor;
