+++ packages/perseus-editor/src/widgets/interaction-editor/line-editor.tsx
