+++ packages/perseus-editor/src/widgets/interaction-editor/movable-point-editor.tsx
