+++ packages/perseus-editor/src/widgets/interaction-editor/point-editor.tsx
